requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.2
psycopg2-binary==2.9.9
bcrypt==4.3.0